+++ src/runtime/parallel/run_slave.js
@@ -1,0 +1,10 @@
+import Slave from './slave'
+
+export default async function run() {
+  const slave = new Slave({
+    stdin: process.stdin,
+    stdout: process.stdout,
+    cwd: process.cwd()
+  })
+  await slave.run()
+}